Move Advanced Statistics from libanki to hook with only small changes to libanki/Stats.java
Move Advanced Statistics from libanki to hook with only small changes to
libanki/Stats.java. Introduced StatsMetaInfo which duplicates lots of
fields from stats.java in order to pass fields from hook to stats.java.
