+++ AnkiDroid/src/main/java/com/ichi2/libanki/Stats.java
@@ -19,10 +19,11 @@ package com.ichi2.libanki;
 
 import android.content.Context;
 import android.database.Cursor;
-import android.database.sqlite.SQLiteDatabase;
 
 import com.ichi2.anki.AnkiDroidApp;
 import com.ichi2.anki.R;
+import com.ichi2.anki.stats.StatsMetaInfo;
+import com.ichi2.libanki.hooks.Hooks;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -72,18 +73,11 @@ public class Stats {
     private double mPeak;
     private double mMcount;
 
-    // JPR new
-    private Settings Settings = new Settings();
-    private ArrayUtils ArrayUtils = new ArrayUtils();
-    private Deck Deck = new Deck();
-    // JPR end
-
     public Stats(Collection col, boolean wholeCollection) {
         mCol = col;
         mWholeCollection = wholeCollection;
         sCurrentInstance = this;
     }
-
 
     public static Stats currentStats() {
         return sCurrentInstance;
@@ -177,6 +171,42 @@ public class Stats {
 
         return new int[]{cards, thetime, failed, lrn, rev, relrn, filt, mcnt, msum};
     }
+
+    //JPR
+
+    public boolean calculateDue(Context context, int type) {
+
+        StatsMetaInfo metaInfo = new StatsMetaInfo();
+
+        metaInfo = (StatsMetaInfo) Hooks.getInstance(context).runFilter("advancedStatistics", metaInfo, type, mCol);
+
+        if(metaInfo.isStatsCalculated()) {
+
+            mDynamicAxis = metaInfo.ismDynamicAxis();
+            mHasColoredCumulative = metaInfo.ismHasColoredCumulative();
+            mType = metaInfo.getmType();
+            mTitle = metaInfo.getmTitle();
+            mBackwards = metaInfo.ismBackwards();
+            mValueLabels = metaInfo.getmValueLabels();
+            mColors = metaInfo.getmColors();
+            mAxisTitles = metaInfo.getmAxisTitles();
+            mMaxCards = metaInfo.getmMaxCards();
+            mMaxElements = metaInfo.getmMaxElements();
+            mFirstElement = metaInfo.getmFirstElement();
+            mLastElement = metaInfo.getmLastElement();
+            mZeroIndex = metaInfo.getmZeroIndex();
+            mCumulative = metaInfo.getmCumulative();
+            mMcount = metaInfo.getmMcount();
+
+            mSeriesList = metaInfo.getmSeriesList();
+
+            return metaInfo.isDataAvailable();
+        }
+        else {
+            return calculateDue(type);
+        }
+    }
+
     /**
      * Due and cumulative due
      * ***********************************************************************************************
@@ -187,78 +217,99 @@ public class Stats {
         mDynamicAxis = true;
         mBackwards = false;
         mTitle = R.string.stats_forecast;
-
-        //JPR
-        //mValueLabels = new int[] { R.string.statistics_young, R.string.statistics_mature };
-        //mColors = new int[] { R.color.stats_young, R.color.stats_mature };
-
-        mValueLabels = new int[] { R.string.statistics_relearn,
-                                   R.string.statistics_mature,
-                                   R.string.statistics_young,
-                                   R.string.statistics_learn};
-        mColors = new int[] {      R.color.stats_relearn,
-                                   R.color.stats_mature,
-                                   R.color.stats_young,
-                                   R.color.stats_learn};
-        //JPR end
-
+        mValueLabels = new int[] { R.string.statistics_young, R.string.statistics_mature };
+        mColors = new int[] { R.color.stats_young, R.color.stats_mature };
         mAxisTitles = new int[] { type, R.string.stats_cards, R.string.stats_cumulative_cards };
+        int end = 0;
+        int chunk = 0;
+        switch (type) {
+            case TYPE_MONTH:
+                end = 31;
+                chunk = 1;
+                break;
+            case TYPE_YEAR:
+                end = 52;
+                chunk = 7;
+                break;
+            case TYPE_LIFE:
+                end = -1;
+                chunk = 30;
+                break;
+        }
 
-        //JPR (moved to own method + replace with new method)
-        //ArrayList<int[]> dues = calculateDues(type);
-        ArrayList<int[]> dues = calculateDuesWithSimulation(type);
+        String lim = "";// AND due - " + mCol.getSched().getToday() + " >= " + start; // leave this out in order to show
+        // card too which were due the days before
+        if (end != -1) {
+            lim += " AND day <= " + end;
+        }
 
-        //mSeriesList = new double[3][dues.size()];
-        mSeriesList = new double[5][dues.size()];
-        //JPR end
+        ArrayList<int[]> dues = new ArrayList<int[]>();
+        Cursor cur = null;
+        try {
+            String query;
+            query = "SELECT (due - " + mCol.getSched().getToday() + ")/" + chunk
+                    + " AS day, " // day
+                    + "count(), " // all cards
+                    + "sum(CASE WHEN ivl >= 21 THEN 1 ELSE 0 END) " // mature cards
+                    + "FROM cards WHERE did IN " + _limit() + " AND queue IN (2,3)" + lim
+                    + " GROUP BY day ORDER BY day";
+            Timber.d("Forecast query: %s", query);
+            cur = mCol
+                    .getDb()
+                    .getDatabase()
+                    .rawQuery(query, null);
+            while (cur.moveToNext()) {
+                dues.add(new int[] { cur.getInt(0), cur.getInt(1), cur.getInt(2) });
+            }
+        } finally {
+            if (cur != null && !cur.isClosed()) {
+                cur.close();
+            }
+        }
+        // small adjustment for a proper chartbuilding with achartengine
+        if (dues.size() == 0 || dues.get(0)[0] > 0) {
+            dues.add(0, new int[] { 0, 0, 0 });
+        }
+        if (end == -1 && dues.size() < 2) {
+            end = 31;
+        }
+        if (type != TYPE_LIFE && dues.get(dues.size() - 1)[0] < end) {
+            dues.add(new int[] { end, 0, 0 });
+        } else if (type == TYPE_LIFE && dues.size() < 2) {
+            dues.add(new int[] { Math.max(12, dues.get(dues.size() - 1)[0] + 1), 0, 0 });
+        }
 
+        mSeriesList = new double[3][dues.size()];
         for (int i = 0; i < dues.size(); i++) {
             int[] data = dues.get(i);
 
-            //JPR
-            //if(data[1] > mMaxCards)
-            //    mMaxCards =data[1];                                           //Y-Axis: Max. value
-            if(data[1] + data[2] + data[3] + data[4] > mMaxCards)
-                mMaxCards = data[1] + data[2] + data[3] + data[4];
+            if(data[1] > mMaxCards)
+                mMaxCards =data[1];
 
-            //mSeriesList[0][i] = data[0];                                      //X-Axis: Day / Week / Month
-            //mSeriesList[1][i] = data[1];                                      //Y-Axis: # Cards
-            //mSeriesList[2][i] = data[2];                                      //Y-Axis: # Mature cards
-
-            //   1 +
-            //   0 = Learn
-            //   1 = Young
-            //   2 = Mature
-            //   3 = Relearn
-            mSeriesList[0][i] = data[0];                                        //X-Axis: Day / Week / Month
-            mSeriesList[1][i] = data[1] + data[2] + data[3] + data[4];          //Y-Axis: # Cards
-            mSeriesList[2][i] = data[1] + data[2] + data[3];                    //Y-Axis: # Mature cards
-            mSeriesList[3][i] = data[1] + data[2];                              //Y-Axis: # Young
-            mSeriesList[4][i] = data[1];                                        //Y-Axis: # Learn
-
-            //JPR end
-
+            mSeriesList[0][i] = data[0];
+            mSeriesList[1][i] = data[1];
+            mSeriesList[2][i] = data[2];
             if(data[0] > mLastElement)
-                mLastElement = data[0];         //X-Axis: Max. value (only for TYPE_LIFE)
+                mLastElement = data[0];
             if(data[0] == 0){
-                mZeroIndex = i;                 //Because we retrieve dues in the past and we should not cumulate them
+                mZeroIndex = i;
             }
         }
-        mMaxElements = dues.size()-1;           //# X values
+        mMaxElements = dues.size()-1;
         switch (mType) {
             case TYPE_MONTH:
-                mLastElement = 31;              //X-Axis: Max. value
+                mLastElement = 31;
                 break;
             case TYPE_YEAR:
-                mLastElement = 52;              //X-Axis: Max. value
+                mLastElement = 52;
                 break;
             default:
         }
-        mFirstElement = 0;                      //X-Axis: Min. value
+        mFirstElement = 0;
 
         mHasColoredCumulative = false;
-        mCumulative = Stats.createCumulative(new double[][]{mSeriesList[0], mSeriesList[1]}, mZeroIndex);   //Day starting at mZeroIndex, Cumulative # cards
-        mMcount = mCumulative[1][mCumulative[1].length-1];                                                  //Y-Axis: Max. cumulative value
+        mCumulative = Stats.createCumulative(new double[][]{mSeriesList[0], mSeriesList[1]}, mZeroIndex);
+        mMcount = mCumulative[1][mCumulative[1].length-1];
         //some adjustments to not crash the chartbuilding with emtpy data
         if(mMaxElements == 0){
             mMaxElements = 10;
@@ -337,652 +388,6 @@ public class Stats {
         }
 
         return dues;
-    }
-
-    private ArrayList<int[]> calculateDuesWithSimulation(int type) {
-        int end = 0;
-        int chunk = 0;
-        switch (type) {
-            case TYPE_MONTH:
-                end = 31;
-                chunk = 1;
-                break;
-            case TYPE_YEAR:
-                end = 52;
-                chunk = 7;
-                break;
-            case TYPE_LIFE:
-                end = -1;
-                chunk = 30;
-                break;
-        }
-
-        ArrayList<int[]> dues = new ArrayList<int[]>();
-
-        EaseClassifier classifier = new EaseClassifier(mCol.getDb().getDatabase());
-        ReviewSimulator reviewSimulator = new ReviewSimulator(mCol.getDb().getDatabase(), classifier);
-
-        SimulationResult simulationResult = reviewSimulator.simNreviews();
-
-        int[][] nReviews = ArrayUtils.transposeMatrix(simulationResult.getNReviews());
-        //int[][] nInState = simulationResult.getNInState();
-
-        // Forecasted number of reviews
-        //   0 = Learn
-        //   1 = Young
-        //   2 = Mature
-        //   3 = Relearn
-
-        for(int i = 0; i<nReviews.length; i++) {
-            dues.add(new int[] { i,  nReviews[i][0], nReviews[i][1], nReviews[i][2], nReviews[i][3] });
-        }
-
-        // small adjustment for a proper chartbuilding with achartengine
-        if (dues.size() == 0 || dues.get(0)[0] > 0) {
-            dues.add(0, new int[] { 0, 0, 0, 0, 0 });
-        }
-        if (end == -1 && dues.size() < 2) {
-            end = 31;
-        }
-        if (type != TYPE_LIFE && dues.get(dues.size() - 1)[0] < end) {
-            dues.add(new int[] { end, 0, 0, 0, 0 });
-        } else if (type == TYPE_LIFE && dues.size() < 2) {
-            dues.add(new int[] { Math.max(12, dues.get(dues.size() - 1)[0] + 1), 0, 0, 0, 0 });
-        }
-
-        return dues;
-    }
-
-    private class Card {
-        private int ivl;
-        private double factor;
-        private int due;
-        private int correct;
-        private long id;
-
-        @Override
-        public String toString() {
-            return "Card [ivl=" + ivl + ", factor=" + factor + ", due=" + due + ", correct=" + correct + ", id="
-                    + id + "]";
-        }
-
-        public Card(long id, int ivl, int factor, int due, int correct) {
-            super();
-            this.id = id;
-            this.ivl = ivl;
-            this.factor = factor / 1000.0;
-            this.due = due;
-            this.correct = correct;
-        }
-
-        public Card clone() {
-            return new Card(id, ivl, (int) (factor * 1000), due, correct);
-        }
-
-        public long getId() {
-            return id;
-        }
-
-        public int getIvl() {
-            return ivl;
-        }
-
-        public void setIvl(int ivl) {
-            this.ivl = ivl;
-        }
-
-        public double getFactor() {
-            return factor;
-        }
-
-        public void setFactor(double factor) {
-            this.factor = factor;
-        }
-
-        public int getDue() {
-            return due;
-        }
-
-        public void setDue(int due) {
-            this.due = due;
-        }
-
-        public int getType() {
-            //# 0=new, 1=Young, 2=mature
-            if(ivl == 0) {
-                return 0;
-            } else if (ivl >= 21) {
-                return 2;
-            } else {
-                return 1;
-            }
-        }
-
-        public int getCorrect() {
-            return correct;
-        }
-
-        public void setCorrect(int correct) {
-            this.correct = correct;
-        }
-    }
-
-    private class Deck {
-
-        public List<Card> deckFromDB(SQLiteDatabase db, int today) {
-            List<Card> deck = new ArrayList<Card>();
-
-            Cursor cur = null;
-            try {
-                String query;
-                query = "SELECT id, due, ivl, factor, type, reps " +
-                        "FROM cards " +
-                        "order by id;";
-                Timber.d("Forecast query: %s", query);
-                cur = db.rawQuery(query, null);
-                while (cur.moveToNext()) {
-
-                    Card card = new Card(cur.getLong(0),                                            //Id
-                                         cur.getInt(5) == 0 ? 0 : cur.getInt(2),  		            //reps = 0 ? 0 : card interval
-                                         cur.getInt(3) > 0 ? cur.getInt(3) :  2500,                 //factor
-                                         Math.max(cur.getInt(1) - today, 0),                        //due
-                                         1);                                                        //correct
-                    deck.add(card);
-                }
-            } finally {
-                if (cur != null && !cur.isClosed()) {
-                    cur.close();
-                }
-            }
-
-            return deck;
-        }
-    }
-
-    private class EaseClassifier {
-
-        private Random random;
-
-        private SQLiteDatabase db;
-        private double[][] probabilitiesCumulative;
-
-        //# Prior that half of new cards are answered correctly
-        private int[] priorNew = {5, 0, 5, 0};		//half of new cards are answered correctly
-        private int[] priorYoung = {1, 0, 9, 0};	//90% of young cards get "good" response
-        private int[] priorMature = {1, 0, 9, 0};	//90% of mature cards get "good" response
-
-        private String queryBaseNew =
-                "select "
-                        +   "count() as N, "
-                        +   "sum(case when ease=1 then 1 else 0 end) as repeat, "
-                        +   "0 as hard, "	//Doesn't occur in query_new
-                        +	  "sum(case when ease=2 then 1 else 0 end) as good, "
-                        +	  "sum(case when ease=3 then 1 else 0 end) as easy "
-                        + "from revlog ";
-
-        private String queryBaseYoungMature =
-                "select "
-                        +   "count() as N, "
-                        +   "sum(case when ease=1 then 1 else 0 end) as repeat, "
-                        +   "sum(case when ease=2 then 1 else 0 end) as hard, "	//Doesn't occur in query_new
-                        +	  "sum(case when ease=3 then 1 else 0 end) as good, "
-                        +	  "sum(case when ease=4 then 1 else 0 end) as easy "
-                        + "from revlog ";
-
-        private String queryNew =
-                queryBaseNew
-                        + "where type=0;";
-
-        private String queryYoung =
-                queryBaseYoungMature
-                        + "where type=1 and lastIvl < 21;";
-
-        private String queryMature =
-                queryBaseYoungMature
-                        + "where type=1 and lastIvl >= 21;";
-
-        public EaseClassifier(SQLiteDatabase db) {
-            this.db = db;
-            this.probabilitiesCumulative = calculateCumProbabilitiesForNewEasePerCurrentEase();
-
-            System.out.println("new\t" + Arrays.toString(this.probabilitiesCumulative[0]));
-            System.out.println("young\t" + Arrays.toString(this.probabilitiesCumulative[1]));
-            System.out.println("mature\t" + Arrays.toString(this.probabilitiesCumulative[2]));
-
-            random = new Random();
-        }
-
-        private double[] cumsum(double[] p) {
-
-            p[1] = p[0] + p[1];
-            p[2] = p[1] + p[2];
-            p[3] = p[2] + p[3];
-
-            return p;
-        }
-
-        private double[][] calculateCumProbabilitiesForNewEasePerCurrentEase() {
-            double[][] p = new double[3][];
-
-            p[0] = cumsum(calculateProbabilitiesForNewEaseForCurrentEase(queryNew, priorNew));
-            p[1] = cumsum(calculateProbabilitiesForNewEaseForCurrentEase(queryYoung, priorYoung));
-            p[2] = cumsum(calculateProbabilitiesForNewEaseForCurrentEase(queryMature, priorMature));
-
-            return p;
-        }
-
-        private double[] calculateProbabilitiesForNewEaseForCurrentEase(String queryNewEaseCountForCurrentEase, int[] prior) {
-
-            Cursor cur = null;
-
-            int[] freqs = new int[] {
-                    prior[0],
-                    prior[1],
-                    prior[2],
-                    prior[3]
-            };
-
-            int nPrior = prior[0] + prior[1] + prior[2] + prior[3];
-
-            int n = nPrior;
-
-            try {
-                cur = db.rawQuery(queryNewEaseCountForCurrentEase, null);
-                cur.moveToNext();
-
-                freqs[0] += cur.getInt(1);          //Repeat
-                freqs[1] += cur.getInt(2);          //Hard
-                freqs[2] += cur.getInt(3);          //Good
-                freqs[3] += cur.getInt(4);          //Easy
-
-                int nQuery = cur.getInt(0);        //N
-
-                n += nQuery;
-
-            } finally {
-                if (cur != null && !cur.isClosed()) {
-                    cur.close();
-                }
-            }
-
-            double[] probs = new double[] {
-                    freqs[0] / (double) n,
-                    freqs[1] / (double) n,
-                    freqs[2] / (double) n,
-                    freqs[3] / (double) n
-            };
-
-            return probs;
-        }
-
-        private int draw(double[] p) {
-            return searchsorted(p, random.nextDouble());
-        }
-
-        private int searchsorted(double[] p, double random) {
-            if(random <= p[0]) return 0;
-            if(random <= p[1]) return 1;
-            if(random <= p[2]) return 2;
-            return 3;
-        }
-
-        public Card simSingleReview(Card c, boolean preserveCard) {
-            if(preserveCard)
-                c = c.clone();
-            return simSingleReview(c);
-        }
-
-        public Card simSingleReview(Card c){
-
-            int type = c.getType();
-
-            int outcome = draw(probabilitiesCumulative[type]);
-
-            applyOutcomeToCard(c, outcome);
-
-            return c;
-        }
-
-        private void applyOutcomeToCard(Card c, int outcome) {
-
-            int type = c.getType();
-            int ivl = c.getIvl();
-            double factor = c.getFactor();
-
-            if(type == 0) {
-                if (outcome <= 2)
-                    ivl = 1;
-                else
-                    ivl = 4;
-            }
-            else {
-                switch(outcome) {
-                    case 0:
-                        ivl = 1;
-                        break;
-                    case 1:
-                        ivl *= 1.2;
-                        break;
-                    case 2:
-                        ivl *= 1.2 * factor;
-                        break;
-                    case 3:
-                    default:
-                        ivl *= 1.2 * 2. * factor;
-                        break;
-                }
-            }
-
-            c.setIvl(ivl);
-            c.setCorrect((outcome > 0) ? 1 : 0);
-            //c.setTypetype);
-            //c.setIvl(60);
-            //c.setFactor(factor);
-        }
-    }
-
-    public class NewCardSimulator {
-
-        private int maxAddPerDay=20;
-
-        private int nAddedToday = 0;
-        private int tAdd = 0;
-
-        public int simulateNewCard(Card card) {
-            nAddedToday++;
-            int tElapsed = tAdd;	//differs from online
-            if (nAddedToday >= maxAddPerDay) {
-                tAdd++;
-                nAddedToday = 0;
-            }
-            return tElapsed;
-        }
-    }
-
-    private class ReviewSimulator {
-
-        private SQLiteDatabase db;
-        private EaseClassifier classifier;
-
-        //TODO: also exists in Review
-        private int nTimeBins = Settings.getNTimeBins();
-        private int timeBinLength = Settings.getTimeBinLength();
-
-        private int tMax = nTimeBins * timeBinLength;
-
-        private NewCardSimulator newCardSimulator = new NewCardSimulator();
-
-        public ReviewSimulator(SQLiteDatabase db, EaseClassifier classifier) {
-            this.db = db;
-            this.classifier = classifier;
-        }
-
-        public SimulationResult simNreviews() {
-
-            SimulationResult simulationResult = new SimulationResult(nTimeBins);
-
-            //nSmooth=1
-
-            //TODO:
-            //Forecasted final state of deck
-            //finalIvl = np.empty((nSmooth, nCards), dtype='f8')
-
-            int currentTime = (int) (System.currentTimeMillis() / 1000);
-
-            int today = (int) ((currentTime - Settings.getDeckCreationTimeStamp()) / Settings.getNSecsPerDay());
-            int mDayCutoff = Settings.getDeckCreationTimeStamp() + ((today + 1) * Settings.getNSecsPerDay());
-
-            System.out.println("today: " + today);
-            System.out.println("todayCutoff: " + mDayCutoff);
-
-            List<Card> deck = Deck.deckFromDB(db, today);
-
-            Stack<Review> reviews = new Stack<Review>();
-
-            //TODO: by having simulateReview add future reviews depending on which simulation of this card this is (the nth) we can:
-            //1. Do monte carlo simulation if we add k future reviews if n = 1
-            //2. Do a complete traversal of the future reviews tree if we add k future reviews for all n
-            //3. Do any combination of these
-            for (Card card : deck) {
-
-                Review review = new Review(card, simulationResult, newCardSimulator, classifier, reviews);
-
-                if (review.getT() < tMax)
-                    reviews.push(review);
-
-                while (!reviews.isEmpty()) {
-                    review = reviews.pop();
-                    review.simulateReview();
-                }
-
-            }
-
-            ArrayUtils.formatMatrix(simulationResult.getNReviews(), "%02d ");
-            ArrayUtils.formatMatrix(simulationResult.getNInState(), "%02d ");
-
-            return simulationResult;
-        }
-    }
-
-    private class Settings {
-        //TODO
-        public int getDeckCreationTimeStamp() {
-            return 1445619600;
-        }
-
-        public int getNTimeBins() {
-            return 30;
-        }
-
-        public int getTimeBinLength() {
-            return 1;
-        }
-
-        public int getMaxReviewsPerDay() {
-            return 10000;
-        }
-
-        public long getNow() {
-            return 1451223980146L;
-            //return System.currentTimeMillis();
-        }
-
-        public int getToday() {
-            int currentTime = (int) (getNow() / 1000);
-            int today = (int) ((currentTime - getDeckCreationTimeStamp()) / getNSecsPerDay());
-            return today;
-        }
-
-        public int getNSecsPerDay() {
-            return 86400;
-        }
-
-        public int getTodayCutoffSecsFromDeckCreation() {
-            int mDayCutoff = getDeckCreationTimeStamp() + ((getToday() + 1) * getNSecsPerDay());
-            return mDayCutoff;
-        }
-    }
-
-    private class ArrayUtils {
-        public int[][] createIntMatrix(int m, int n) {
-            int[][] matrix = new int[m][];
-            for(int i=0; i<m; i++) {
-                matrix[i] = new int[n];
-                for(int j=0; j<n; j++)
-                    matrix[i][j] = 0;
-            }
-
-            return matrix;
-        }
-
-        public int[][] transposeMatrix(int[][] matrix) {
-            if (matrix.length == 0)
-                return matrix;
-
-            int m = matrix.length;
-            int n = matrix[0].length;
-
-            int transpose[][] = new int[n][m];
-
-            int c, d;
-            for ( c = 0 ; c < m ; c++ )
-            {
-                for ( d = 0 ; d < n ; d++ )
-                    transpose[d][c] = matrix[c][d];
-            }
-
-            return transpose;
-        }
-
-        public void formatMatrix(int[][] matrix, String format) {
-            StringBuilder s = new StringBuilder();
-
-            for(int i=0; i<matrix.length; i++) {
-                for(int j=0; j<matrix[i].length; j++) {
-                    s.append(String.format(format, matrix[i][j]));
-                    s.append(System.getProperty("line.separator"));
-                }
-            }
-
-            Timber.d(s.toString());
-        }
-    }
-
-    private class SimulationResult {
-
-        private int timeBinLength = Settings.getTimeBinLength();
-        private int tMax;
-
-        // Forecasted number of reviews
-        //   0 = Learn
-        //   1 = Young
-        //   2 = Mature
-        //   3 = Relearn
-        private int[][] nReviews;
-
-        // Forecasted number of cards per state
-        //   0 = New
-        //   1 = Young
-        //   2 = Mature
-        private int[][] nInState;
-
-        public SimulationResult(int nTimeBins) {
-            nReviews = ArrayUtils.createIntMatrix(4, nTimeBins);
-            nInState = ArrayUtils.createIntMatrix(3, nTimeBins);
-
-            tMax = nTimeBins * timeBinLength;
-        }
-
-        public SimulationResult(int[][] nReviews, int[][] nInState) {
-            this.nReviews = nReviews;
-            this.nInState = nInState;
-        }
-
-        public int[][] getNReviews() {
-            return nReviews;
-        }
-
-        public int[][] getNInState() {
-            return nInState;
-        }
-
-        public int nReviewsDoneToday(int tElapsed) {
-            //This excludes new cards and relearns
-            return nReviews[1][tElapsed / timeBinLength] +
-                    nReviews[2][tElapsed / timeBinLength];
-        }
-
-        public void incrementNInState(int cardType, int t) {
-            nInState[cardType][t]++;
-        }
-
-        public void incrementNReviews(int cardType, int t) {
-            nReviews[cardType][t]++;
-        }
-
-        public void updateNInState(Card card, int tFrom, int tTo) {
-            for(int t = tFrom / timeBinLength; t < tTo / timeBinLength; t++)
-                if(t < tMax) {
-                    nInState[card.getType()][t]++;
-                }
-        }
-    }
-
-    private class Review {
-
-        private int maxReviewsPerDay = Settings.getMaxReviewsPerDay();
-
-        private int nTimeBins = Settings.getNTimeBins();
-        private int timeBinLength = Settings.getTimeBinLength();
-
-        private int tMax = nTimeBins * timeBinLength;
-
-        private int tElapsed;
-        private Card card;
-        private SimulationResult simulationResult;
-        private EaseClassifier classifier;
-        private Stack<Review> reviews;
-
-        private Review (Card card, SimulationResult simulationResult, EaseClassifier classifier, Stack<Review> reviews, int tElapsed) {
-            this.card = card;
-            this.simulationResult = simulationResult;
-            this.classifier = classifier;
-            this.reviews = reviews;
-
-            this.tElapsed = tElapsed;
-        }
-
-        public Review (Card card, SimulationResult simulationResult, NewCardSimulator newCardSimulator, EaseClassifier classifier, Stack<Review> reviews) {
-            this.card = card;
-            this.simulationResult = simulationResult;
-            this.classifier = classifier;
-            this.reviews = reviews;
-
-            //# Rate-limit new cards by shifting starting time
-            if (card.getType() == 0)
-                tElapsed = newCardSimulator.simulateNewCard(card);
-            else
-                tElapsed = card.getDue();
-
-            // Set state of card between start and first review
-            this.simulationResult.updateNInState(card, 0, tElapsed);
-        }
-
-        public void simulateReview() {
-            //Set state of card for current review
-            simulationResult.incrementNInState(card.getType(), tElapsed / timeBinLength);
-
-            if(card.getType() == 0 || simulationResult.nReviewsDoneToday(tElapsed) < maxReviewsPerDay) {
-                // Update the forecasted number of reviews
-                simulationResult.incrementNReviews(card.getType(), tElapsed / timeBinLength);
-
-                // Simulate response
-                card = classifier.simSingleReview(card, true);
-
-                // If card failed, update "relearn" count
-                if(card.getCorrect() == 0)
-                    simulationResult.incrementNReviews(3, tElapsed / timeBinLength);
-
-                // Set state of card between current and next review
-                simulationResult.updateNInState(card, tElapsed + 1, tElapsed + card.getIvl());
-
-                // Advance time to next review
-                tElapsed += card.getIvl();
-            }
-            else {
-                // Advance time to next review (max. #reviews reached for this day)
-                tElapsed += 1;
-            }
-
-            if (tElapsed < tMax) {
-                Review review = new Review(card, simulationResult, classifier, reviews, tElapsed);
-                this.reviews.push(review);
-            }
-        }
-
-        public int getT() {
-            return tElapsed;
-        }
     }
 
     //30-12-2015 JPR new end
